Changed balance entity to Account, now Users can have multiple accounts, implemented repository and controller for it, added some javadoc

+++ src/main/java/hu/csercsak_albert/banking_system/dto/TransactionDto.java
@@ -21,9 +21,9 @@ public class TransactionDto {
 
     private TransactionType transactionType;
 
-    private int fromAccountNumber;
+    private long fromAccountNumber;
 
-    private int toAccountNumber;
+    private long toAccountNumber;
 
     @Min(1)
     private double amount;
